Split tests.py into specific test cases
